+++ rocketpy/__init__.py
@@ -30,6 +30,9 @@ from .rocket import (
     Components,
     EllipticalFins,
     Fins,
+    FreeFormFins,
+    GenericSurface,
+    LinearGenericSurface,
     NoseCone,
     Parachute,
     RailButtons,
@@ -37,6 +40,8 @@ from .rocket import (
     Tail,
     TrapezoidalFins,
 )
+from .sensitivity import SensitivityModel
+from .sensors import Accelerometer, Barometer, GnssReceiver, Gyroscope
 from .simulation import Flight, MonteCarlo
 from .stochastic import (
     StochasticEllipticalFins,