+++ tests/unit/test_environment.py
@@ -71,13 +71,18 @@ def test_location_set_topographic_profile_computes_elevation(
 
 def test_geodesic_coordinate_geodesic_to_utm_converts_coordinate():
     """Tests the conversion from geodesic to UTM coordinates."""
-    x, y, utm_zone, utm_letter, north_south_hemis, east_west_hemis = (
-        Environment.geodesic_to_utm(
-            lat=32.990254,
-            lon=-106.974998,
-            semi_major_axis=6378137.0,  # WGS84
-            flattening=1 / 298.257223563,  # WGS84
-        )
+    (
+        x,
+        y,
+        utm_zone,
+        utm_letter,
+        north_south_hemis,
+        east_west_hemis,
+    ) = Environment.geodesic_to_utm(
+        lat=32.990254,
+        lon=-106.974998,
+        semi_major_axis=6378137.0,  # WGS84
+        flattening=1 / 298.257223563,  # WGS84
     )
     assert np.isclose(x, 315468.64, atol=1e-5)
     assert np.isclose(y, 3651938.65, atol=1e-5)